server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
openpyxl==3.1.2
python-dotenv==1.0.0
werkzeug==2.3.7
pyyaml==6.0.1
sqlalchemy==2.0.25
pytest==7.4.4
python-dateutil==2.8.2
apscheduler==3.10.4